+++ src/senaite/storage/indexers/storage_layout_container.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 
-from bika.lims.catalog.indexers import get_searchable_text_tokens
 from plone.indexer import indexer
-from senaite.storage.catalog import SENAITE_STORAGE_CATALOG
+from senaite.core.catalog.utils import get_searchable_text_tokens
+from senaite.storage.catalog import STORAGE_CATALOG
 from senaite.storage.interfaces import ISenaiteStorageCatalog
 from senaite.storage.interfaces import IStorageLayoutContainer
 
@@ -12,7 +12,7 @@ def listing_searchable_text(instance):
     """Add searchable text tokens for storage containers
     """
     entries = set()
-    tokens = get_searchable_text_tokens(instance, SENAITE_STORAGE_CATALOG)
+    tokens = get_searchable_text_tokens(instance, STORAGE_CATALOG)
     entries.update(tokens)
     entries.update(instance.get_all_ids())
     return u" ".join(list(entries))